minimizing files and put all in one
