+++ web/src/pages/success/index.tsx
@@ -1,7 +1,7 @@
 import { GetServerSideProps } from 'next'
 import { useRouter } from 'next/router'
-import { fetchStripeLineItems } from '@utils/fetchStripeLineItems'
-import { fetchStripeCustomerDetails } from '@utils/fetchStripeCustomerDetails'
+import { fetchStripeLineItems } from '@utils/stripeAPI'
+import { fetchStripeCustomerDetails } from '@utils/stripeAPI'
 import Headline from '@components/Headline'
 import Divider from '@components/Divider'
 import Text from '@components/Text'
