--- web/src/utils/fetchProductById.ts
@@ -1,15 +1,0 @@
-export const fetchProductById = async (slug: string): Promise<IProduct> => {
-  const query = encodeURIComponent(
-    `*[_type == "products" && _id == "${slug}"][0]{
-      ...,
-      categories[]->{
-        slug
-      }
-    }`
-  )
-  const url = `${process.env.SANITY_API_URL}query=${query}`
-  const data = await fetch(url).then((res) => res.json())
-  const product: IProduct = data.result
-
-  return product
-}
