--- web/src/utils/fetchStripeCustomerDetails.ts
@@ -1,12 +1,0 @@
-export const fetchStripeCustomerDetails = async (sessionId: string) => {
-  const res = await fetch(
-    `${process.env.NEXT_PUBLIC_BASE_URL}/api/getStripeSession?session_id=${sessionId}`
-  )
-
-  if (!res.ok) return
-
-  const data = await res.json()
-  const customer = data.customer.customer_details
-
-  return customer
-}
