--- web/src/utils/getStripe.ts
@@ -1,11 +1,0 @@
-import { Stripe, loadStripe } from '@stripe/stripe-js'
-
-let stripePromise: Promise<Stripe | null>
-const getStripe = () => {
-  if (!stripePromise) {
-    stripePromise = loadStripe(process.env.NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY!)
-  }
-  return stripePromise
-}
-
-export default getStripe
